@@ -13,52 +13,52 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~8.1.2",
-    "@angular/compiler": "~8.1.2",
-    "@angular/core": "~8.1.2",
-    "@angular/forms": "~8.1.2",
-    "@angular/platform-browser": "~8.1.2",
-    "@angular/platform-browser-dynamic": "~8.1.2",
-    "@angular/router": "~8.1.2",
-    "@ionic-native/core": "^5.0.0",
-    "@ionic-native/file": "^5.15.0",
-    "@ionic-native/social-sharing": "^5.15.0",
-    "@ionic-native/splash-screen": "^5.0.0",
-    "@ionic-native/status-bar": "^5.0.0",
-    "@ionic/angular": "^4.7.1",
+    "@angular/common": "~10.0.5",
+    "@angular/compiler": "~10.0.5",
+    "@angular/core": "~10.0.5",
+    "@angular/forms": "~10.0.5",
+    "@angular/platform-browser": "~10.0.5",
+    "@angular/platform-browser-dynamic": "~10.0.5",
+    "@angular/router": "~10.0.5",
+    "@ionic-native/core": "^5.27.0",
+    "@ionic-native/file": "^5.27.0",
+    "@ionic-native/social-sharing": "^5.27.0",
+    "@ionic-native/splash-screen": "^5.27.0",
+    "@ionic-native/status-bar": "^5.27.0",
+    "@ionic/angular": "^5.3.0",
     "cordova-plugin-file": "^6.0.2",
-    "cordova-plugin-x-socialsharing": "^5.6.0",
-    "core-js": "^2.5.4",
+    "cordova-plugin-x-socialsharing": "^5.6.8",
+    "core-js": "^3.6.5",
     "ngx-papaparse": "^3.0.4",
-    "rxjs": "~6.5.1",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.9.1"
+    "rxjs": "~6.6.0",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.801.2",
-    "@angular-devkit/build-angular": "~0.801.2",
-    "@angular-devkit/core": "~8.1.2",
-    "@angular-devkit/schematics": "~8.1.2",
-    "@angular/cli": "~8.1.2",
-    "@angular/compiler": "~8.1.2",
-    "@angular/compiler-cli": "~8.1.2",
-    "@angular/language-service": "~8.1.2",
-    "@ionic/angular-toolkit": "~2.0.0",
-    "@types/jasmine": "~3.3.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "^5.0.0",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.4.3"
+    "@angular-devkit/architect": "~0.1000.4",
+    "@angular-devkit/build-angular": "~0.1000.4",
+    "@angular-devkit/core": "~10.0.4",
+    "@angular-devkit/schematics": "~10.0.4",
+    "@angular/cli": "~10.0.4",
+    "@angular/compiler": "~10.0.5",
+    "@angular/compiler-cli": "~10.0.5",
+    "@angular/language-service": "~10.0.5",
+    "@ionic/angular-toolkit": "~2.3.0",
+    "@types/jasmine": "~3.5.11",
+    "@types/jasminewd2": "~2.0.8",
+    "@types/node": "~14.0.25",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.2",
+    "karma": "~5.1.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~3.3.1",
+    "karma-jasmine-html-reporter": "^1.5.4",
+    "protractor": "~7.0.0",
+    "ts-node": "~8.10.2",
+    "tslint": "~6.1.2",
+    "typescript": "~3.9.7"
   },
   "description": "An Ionic project",
   "cordova": {
